Tests/LibWeb/Text/expected/DOM/range-get-client-rects.txt

[object DOMRectList]
[object DOMRect]
PASS
